lightrec: Allow to specify custom I/O ops in the memory map
The current generators for I/O opcodes work fine when the memory area is
a standard RAM or ROM area. However, they don't work as soon as the
memory area requires more complex emulation mechanisms to happen, like
when it corresponds to a memory-mapped set of hardware registers.

This commit introduces a new internal function, that will later be
called from C for every I/O operation.