Make the I/O functions call a C procedure to perform the operation
This is done in order to allow custom I/O functions on certain memory
area. It is required to properly emulate the hardware when the target of
an operation is a hardware register.

The old I/O functions have not been removed, but marked as "meta"
instead; they will be reused at some point, when we can determine at
compilation time that the address targetted by the I/O operation
corresponds to a standard memory area (e.g. RAM or ROM).