Add first bits of code invalidation mechanism
Through the function lightrec_invalidate(), Lightrec will handle
invalidation of all the blocks which overlap the memory area specified
by the function parameters, if the memory map entry it corresponds to
has the MAP_IS_RWX flag set in its newly added 'flags' field.

Right now, this commit only adds a way to specify that a memory area is
invalidated - the recompiled blocks are not yet recompiled
automatically.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>