Enhance the mechanism of invalidation
Detecting that a block is outdated by checking its compilation date vs.
the modification date of the memory pages is broken, at least if the
cycles counter is only 32 bits. For instance, if the emulated CPU runs
at a nominal frequency of 32 MHz, the timer would overflow in less than
three minutes.

To fix that, and avoid switching the cycles counter to a 64-bits value,
we store a "hash" of the memory in the block structure, when the block
is compiled. This "hash" corresponds to the sum of the cycles values of
the pages that the block covers. Now, to detect that a block is
obsolete, we just need to re-calculate this hash and compare it to the
one stored in the block structure. If they don't match, the block is
obsolete.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>