Add a first-pass interpreter step
Instead of being fully recompiled, the blocks are now pre-processed
(to create the opcode list), interpreted, and then finally compiled.

The point of this, is that the interpreter step can generate profile
data (e.g. whether a store operation hit the RAM or hardware registers)
which will then be used by the recompiler to produce optimized code.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>