Add threaded recompiler
Add an option to compile the blocks in a separate thread. This should be
a big improvement even on single-core CPUs, as the VBLANK time (when the
emulator is waiting for the vertical sync) can now be used for
meaningful processing.

Since the threaded recompiler uses C11 threads, the C standard
requirement has been bumped to C11.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>