Optimizer: Fix register unloaded too early
In the case where the one register is written, then read, then written
again, the optimizer would have emitted a unload meta-opcode after both
the opcode reading the register, and the opcode that lasts writes the
register.

We can improve this by emitting the unload meta-opcode after the opcode
reading the register only if it hasn't been written to.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>