Add direct-I/O path for stores
Generate optimized code in the case where a store is aimed at the RAM or
scratchpad, instead of calling into the slow C function.

If the target of the store is the RAM, the code LUT at the corresponding
offset will be zeroed to invalidate any block present at the target of
the write.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>