Support MFC opcodes in branch delay slot
When a branch has a MFC0, CFC0, MFC2, CFC2 or LWC2 opcode in a delay
slot, and the register written to is read by the first opcode of the
block at the target PC, then the old value should be read.

Implement this in the interpreter by emulating the first opcode of the
next block before emulating the opcode, should it be one of the
problematic opcodes listed above.

Fixes character animations in Tekken 2.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>