Avoid using code inside #ifdef
The ENABLE_FIRST_PASS and ENABLE_THREADED_COMPILER option macros now
resolve to 0 or 1, if the option is disabled or enabled respectively.

This permits the code to lie outside of #ifdef blocks, so that the
compiler will compile it unconditionally, allowing to detect syntax
errors even if these options are not enabled. The compiler will
correctly detect the dead codepaths and remove them.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>