interpreter: Handle delay reads only if next opcode reads dest register
When encountering a MFC0/CFC0/MFC2/CFC2/LWC2 opcode in a branch's delay
slot, the destination register of the opcode will be written after the
first opcode of the next block will finish its execution. For that
specific case we always rely on the interpreter.

Before, we unconditionally executed the first opcode of the next block,
which lead to errors as that opcode can be e.g. a branch. Now we do
check that the opcode does read or write the delay slot's destination
register.

Fixes Crash Team Racing's never ending loading screens.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>