CMake: unconditionally enable disassembler on Debug log level
It's bad practice to enable an option if a dependency is found, since
the default build options wouldn't then necesarily be the same.

Now, the disassembler is unconditionally enabled on the Debug log level,
and only at that level. If the libopcodes dependency is no found, an
error message is then displayed.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>