Introduce a 'union code' and use it when calling into wrappers
The 'union code' is the exact same as the transparent union found in the
'struct opcode'.

Most of the current code that manipulates the opcode structure doesn't
need to change the opcode's flags or walk the opcode list. By just
passing the 'union code' instead of 'struct opcode' pointers to the
wrappers.

Since we don't pass pointers to the opcode chain anymore, this opens the
possibility for the opcode chains to be freed as soon as a block has
been compiled.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>