Interpreter: Don't compile next block just to emulate the first opcode
When an opcode in a delay slot performs a delay read, we have to emulate
the first opcode at the target of the jump.

Instead of trying to get a 'struct block *' at the address of the jump,
which will most likely cause a new block to be created, we simply read
one 32-bit opcode and emulate it. The first benefit is then that we
won't recompile a block that will sit unused afterwards.

The second benefit isn't very obvious, but the old code was the only
remaining reason why a block's opcode_list couldn't be freed after the
block has been compiled. This change opens the way to freeing the unused
opcode lists.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>