Blockcache: Reduce size of block LUT
The previous size of 0x40000 was not backed by any test or benchmark. It
resulted in only ~5% of the table being filled.

Reduce it to 0x4000. This results in the table being filled at 60-80%,
while still having an acceptable average number of blocks per entry with
1.5, and an acceptable maximum number of blocks which seems to be lower
than 8.

This reduces the RAM usage of the block cache from 2 MiB to 128 KiB on
64-bit systems and from 1 MiB to 64 KiB on 32-bit systems.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>