Lightrec: Add generic RW wrapper
The difference between the generic RW wrapper and the regular RW wrapper
is that the former takes a 'struct opcode *' and 'struct block *' as
parameters, which means that it can only be used while the IR is still
available. It will pass the 'flags' field of the opcode structure to the
lightrec_rw() function, which will then tag the opcode appropriately.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>