Blockcache: Remove tiny LUT
The tiny LUT was made on the assumption that loading from a tiny LUT
would speed up finding a block because of the reduced dcache pressure.
That's something that I never benchmarked, of course, and now that we
have the code LUT we don't read the block LUT as much as we used to.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>