Lightrec: Add API function lightrec_set_invalidate_mode()
This function can be used to set the invalidate mode.
If the 'dma_only' parameter is false, then the code at the target
address of a CPU write or DMA write is invalidated. If 'dma_only' is
true, only the code at the target address of a DMA write is invalidated.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>