Interpreter: Support return from exception handler to GTE opcode
When an IRQ happens, the PSX exception handlers (when done) will jump
back to the instruction that was executed right before the IRQ, unless
it was a GTE opcode; in that case, it jumps to the instruction right
after. Since we will never handle the IRQ right after a GTE opcode, but
on branch boundaries, we need to adjust the return address so that the
GTE opcode is effectively executed.

This fixes graphical glitches in some games, like Spyro the Dragon.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>