Blockcache: Check block hash if marked as outdated
In many cases, when a block has been invalidated by a I/O write, the
value written is the same as the previous value and the block doesn't
really have to be recompiled. Detect these by calculating a hash of the
memory area covered by the block and comparing the result with the hash
computed when the block was first created.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>