Lightrec: Remove code LUT entries when recompiling a block, not before
Between the moment where a block is tagged as "should recompile" and the
moment where it's effectively recompiled, entries might be added in the
LUT if the block is being compiled in parallel to its interpretation.

To fix that issue, the LUT entries are now removed at the moment where
the block is recompiled, and not just when the block is tagged.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>