Lightrec: Keep old code until a block is recompiled
Before, when a block had to be recompiled, its code was trashed and
unregistered from the code LUT, and the interpreter would execute the
block if needed, until the recompilation is done.

Now, the old code is kept around until it's replaced by the new code,
and the old code is then freed.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>