Recompiler: Fix lockup when shutting down
Prior to this commit, if lightrec_free_recompiler() was called while a
block was being compiled, the compile thread was signaled at a moment
where it was not waiting on the mutex. This is fine, as long as the
"stop" variable is checked before the thread starts to wait, which it
did not do.

Change the thread loop so that the "stop" variable is checked everytime
before waiting on the condition variable.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>