Disassembler: Don't stop block after BREAK opcode
Reaching a BREAK opcode will cause the PS1 to execute a BIOS routine
which will effectively crash the hardware.

With the assumption that PS1 games never trigger this crash, it is
therefore safe to assume that the opcode is never executed, and that
a local branch will jump above it.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>