Optimizer: Remove GCC-generated zero-check sequence after DIV(U)
PSX games compiled with GCC (the majority of them) will have after each
DIV/DIVU opcode a routine that will verify that the divisor is not zero,
and crash otherwise. For signed DIV, it also checks that the result
won't overflow.

Since these code paths never run (as the call to BREAK would crash the
console), they can safely be removed.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>