Emitter: Don't generate zero-check sequence if GCC generated one
If GCC, when compiling the PS1 game, added a zero-check sequence after
the DIV or DIVU opcode, then it is safe to assume that the divisor is
never zero, as this would effectively crash the hardware.

As such, for the DIV/DIVU opcodes that had the GCC-generated zero-check
sequence, we don't have to generate code to handle the case where the
divisor is zero, since we know it will never happen.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>