emitter: Handle sign-extension on output of SRL
The SRL code generator uses a zero-extended input register. Since it's
a right-shift, the upper 32 bits are never changed and the output
register will be zero-extended as well. However, we can also deduce that
if we are shifting by at least one byte, then the bit 31 will be zero,
and therefore the output register will also be sign-extended.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>