emitter: Fix Lightning using live registers in DIV generator
The DIV generator checks that the divisor is zero, and if it's the case,
it branches to a special block of code to handle it.

The branch didn't seem to please Lightning, as it started to use live
registers as its temporaries to generate the DIV opcode.

This fixes Wile E. Coyote falling through the ground of the opening room
in Looney Tunes Sheep Raider.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>