emitter: Mark live registers before jump in DIV generator
Lightning considers that all registers that have not been used between a
branch and jump opcodes are dead. Therefore, we must call
lightrec_regcache_mark_live() before the jump opcode to make sure that
Lightning won't use one of our live registers as a temporary.

This fixes Spyro the Dragon falling through the floor in the demo.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>