disassembler: Add brand new disassembler
This new disassembler does not depend on libopcodes.

This new disassembler will print the original block on the left, then
the modified opcodes (if any) in the center, then the flags on the
right. This makes it much easier to see the results of the optimization
passes.

The other big advantage is that when full debug output is used
(disassembly of original MIPS blocks + disassembly of the generated code
by GNU Lightning), the installed libopcodes only has to support the
host's instruction set.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>