emitter: Get opcode pointer and flags from block and offset
Get a pointer to the current opcode and its flags from the block pointer
and the offset value.

All generator functions will then be able to be simplified, as neither
the 'op' nor the 'pc' arguments are needed to be passed, as they can be
easily recomputed from the block and offset.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>