blockcache: Clear code LUT with a simple memset()
For some time now, when a LUT entry for a block is NULL, Lightrec will
verify whether or not the block is outdated, and will re-insert the LUT
entry if the block is still valid.

Therefore, we can just erase the entire area of the code LUT that's
covered by the block to be unregistered, and Lightrec will recuperate
just fine.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>