emitter: Simplify calling of generic RW wrapper
Since the PC of the most recent entry point is already stored in the
lightrec_state structure, we only need to know the current block's LUT
entry and the opcode's offset to be able to retrieve both the block and
opcode pointers.

Since the LUT offset and opcode offset are both 16-bit values, they can
be passed as one single 32-bit argument, which simplifies the emitted
code even further.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>