optimizer: Don't remove DIV(U) opcode when LO/HI are unused
When the LO/HI registers are never read after the DIV(U) opcode, we
might think it's a good idea to remove the DIV(U) opcode. However this
is not safe to do, as in some games, the LO/HI are read in the parent
function.

This fixes the garbled screen in Tony Hawk Pro Skater 2.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>