Optimizer: Add pass to transform SLL+SRA into 8/16-bit cast
This optimization pass will detect when a SRA follows a SLL with an
immediate value of #24 or #16, and convert them into a EXTC or EXTS
respectively.

It will also try to see if there is a LBU or LHU before the cast, and if
it's the case, convert that opcode to a LB or LH, removing the need for
a cast in the first place.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>