Remove BEQZ/BNEZ meta-opcodes
These meta-opcode didn't really have a purpose, as they were just
representing a BEQ/BNE with one of the registers being zero.

Since the number of meta-opcodes we can implement is limited, remove
these right now.

This will have no effect on the emitted code, since the emitter will
detect when the Rt register is $zero, and emit optimized code in that
case.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>