disassembler: Add COMMIT meta-opcode
Add support for a new COMMIT meta-opcode. If Rd == !register[Rs],
all shadow registers will be copied back to their respective original
registers, and later invalidated. The cycles count can then be adjusted
according to the number of opcodes present in the Imm field.

Otherwise, the shadow registers are not copied, and are all invalidated.
The cycles count is not modified.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>