emitter: Slightly optimize rec_store_direct
When the RAM mirrors are mapped, we don't have to convert the target
address to avoid hitting the mirrors.

This means that instead of using a mask of 0x1fbfffff to only keep the
bits that we're interested in, we can use a mask of 0x1fffffff. The
advantage is that on some archs like MIPS, the latter can be applied in
one opcode (EXT) instead of three (LUI+ORI+AND).

Signed-off-by: Paul Cercueil <paul@crapouillou.net>