emitter: Generate optimized code for coprocessor #0 opcodes
Instead of handling coprocessor #0 opcodes in C, generate code that will
handle reading and writing to the coprocessor #0 registers, and only
jump to C in the particular case where we're writing the status (r12)
register from a function located in the BIOS, as this may toggle the
"RAM enabled" flag.

When writing the status (r12) or cause (r13) registers, the emitted code
will check if a software interrupt (SWI) has been raised, and in this
case, will set the LIGHTREC_EXIT_CHECK_INTERRUPT flag.

Note that it only needs to check for SWIs even if we unmask all
interrupts, because we know that the next interrupt won't happen before
the target cycle.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>