lightrec: Update lightrec_emit_eob() for later use
The function lightrec_emit_eob() would generate a end-of-block and
adjust the cycles counter without taking the current opcode into
consideration (as it is currently used when the current opcode is an
impossible branch).

Update the lightrec_emit_eob() function to add an additional parameter
that specifies whether or not the current opcode should be considered,
which will be useful in the future.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>