lightrec: Fix race between generic RW callback and recompiler
The generic RW callback will attempt to obtain from the block LUT a
pointer to the current block from its PC.

This means that the current block's entry in the block LUT cannot be
removed directly in the lightrec_compile_block() function, as this
function can be executed from the recompiler thread.

Instead, remove it from the block LUT when reaping the block.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>