recompiler: Remove "current_block" pointer
Previously, the block currently being compiled was left on the queue,
but the recompiler would keep a pointer to it.

To get the recompiler to use multiple worker threads, we cannot really
do that; each worker thread would have to check with all other workers
if the block it wants to compile isn't being compiled already, which
really doesn't sound practical.

It wouldn't be a good idea either to dequeue each block before
compiling, because then the lightrec_remove_block() would have no way of
knowing if a block is being compiled or not.

Solve this problem by having a "compiling" boolean in the list entry.
The workers now compile the first element on the list where this
boolean is false, then eventually remove the block from the list.

In lightrec_remove_block(), if we detect that the requested block is
being compiled, simply wait for the worker's notification until the
block has been removed from the list.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>