lightrec: Fix race between multiple compilers, try #2
When more than one compiler thread is running, we must make sure that
the blocks that are going to be replaced by another block's entry
points won't override these entry points before being replaced,
otherwise we have invalid pointers in the code LUT.

Also, we must make sure that the reaper is not running when we reap the
blocks that we are replacing, because there might be duplicates. The
reaper takes care of dropping duplicates, only if it's paused -
otherwise it would maybe end up reaping the same block more than once,
causing double-free errors.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>