optimizer: Fix constant propagation ignored for NOPs
In the "transform ops" step, when a NOP is found the constant
propagation was skipped, with the idea that a NOP does not change any
register.

However the NOPs can be synchronization points. If it's the case, all
known constants (except $zero) must be invalidated.

This fixes Ritcher hovering in the air after jumping in Castlevania -
Symphony of the Night.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>