optimizer: Use kuseg address in LUI if possible
The following pattern is very common in PSX code:

lui $reg, imm
lw $reg, offt($reg)

By fixing imm so that it is known to be in kuseg, the following load
opcode then does not have to mask the input register. The $reg register
temporarily does not hold the correct value, but it doesn't matter as it
is overwritten by the load opcode.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>