optimizer: Rework code updating LUI values
The previous code was a bit broken, as it did not handle the case where
a register was used on both the Rs and Rt fields of a store. This case
was thanksfully extremely rare.

The new code is also better, because it converts a few more LUIs that
weren't converted before, close to SYNC flags.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>