lightrec: Add optional support for external code buffer
Add support for providing a code buffer to Lightrec.

Instead of having GNU Lightning allocate and mmap a buffer for each and
every block generated, the code buffer can be allocated before the
dynarec is started. Lightning then only needs to write into this buffer,
which should greatly reduce the overall number of system calls.

The other benefit is that we have total control on the amount of code
generated: the dynarec will never generate code outside the
predeterminated buffer. Right now Lightrec isn't good at reclaiming old
blocks, so a small code buffer (below 8 MiB) will likely crash the
emulator after a while; but this is a first step in the right direction.

The allocations of code blocks within the code buffers are handled by
the tlsf library (https://github.com/mattconte/tlsf).

Since the tlsf library is not thread-safe, for now the code buffer
support is guarded behind a CMake option, which can only be enabled if
the threaded compiler is disabled. As this code is still pretty
experimental, the new ENABLE_CODE_BUFFER option is disabled by default.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>