Add support for 32-bit code LUT on 64-bit archs
When using an external code buffer whose start and end addresses fit in
32 bits, then all functions that the dynarec will generate will have
32-bit addresses. In this case, using 64-bit addresses in the code LUT
means wasting half of the data space, since each 64-bit entry will
always have its upper 32 bits zeroed.

When it is detected that the code buffer's start and end addresses fit
in 32 bits, each pointer in the code LUT is now 32 bits, and the size of
the code LUT can be reduced by half.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>