Emitter: Fix unaligned access to code LUT
A store opcode can have an immediate operand that is not aligned to the
size of the memory operation, as long as the sum of the source register
with the immediate operand is aligned.

If the immediate operand is not aligned to 4 bytes (since we clear the
low 2 bits of the source register), we fall back to adding it to the
source register before performing the code invalidation; otherwise, it
is possible that the invalidation will hit a different entry in the code
LUT, or even do an unaligned memory access, which is a fault on many
architectures.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>