emitter: Support big-endian in CP2 MTC/MFC/CTC/CFC opcodes
Some CP2 registers are 16-bits signed, some are 16-bits unsigned, some
are 32-bits.

However, the array in the lightrec state structure that contains the
value of the CP2 registers is an array of 32-bit values. The emitter
then generates 16-bit loads/stores to access the 16-bit registers from
this array, using the same offset as if it was reading a 32-bit word.

This meant it only worked on little-endian systems, because on
big-endian system it would access the high 16 bits of the 32-bit
storage, which consists of padding.

Therefore, on big-endian systems the offset to the CP2 register storage
array must be updated so that the 16-bit access will address the low 16
bits of the 32-bit word.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>