Remove TinyMM support
The TinyMM was deemed useful when the IR was a linked list of
individually allocated 'struct opcode' objects. Now, the opcode list is
actually an array, and it's allocated at once. Besides, TinyMM was added
with the assumption that it would be better than calling malloc() for
each and every opcode, but I never benchmarked it, so it's not even sure
that it was doing anything benefical.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>