blockcache: Add function lightrec_remove_outdated_blocks()
This function can be called to remove all blocks detected as outdated in
the block cache, safe for the block optionally passed in the 'except'
parameter.

Very old blocks that aren't used anymore might not be detected as
outdated and will clobber the cache. Since this function is designed to
be called when running low on code cache memory, it will also drop all
the blocks that have been compiled more than ~4 seconds ago. This may
include many blocks that are still used, but those can simply be
recompiled.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>