Handle code buffer overflowing
If the threaded compiler is disabled and a code allocation fails, call
lightrec_remove_outdated_blocks() then try again.

If the threaded compiler is enabled and a code allocation fails, all the
jobs in the recompiler's queue will be flushed, and the recompiler will
request the code reaper to run lightrec_remove_outdated_blocks(). Until
it happens, the input queue is locked, so that no compilation requests
can be enqueued.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>