lightrec: Fix lightrec_block_is_outdated()
The lightrec_block_is_outdated() would check the code LUT at the address
that corresponds to the block's main entry point. Or at least, that's
what it was supposed to do. It was calling lut_read() with the offset
converted using lut_offset(); but lut_read() would convert the offset
itself, causing the offset to be converted twice to an invalid value.

Address this problem by removing the lut_offset() inside lut_read(),
leaving it to the calling functions, to make it similar to lut_write().

Signed-off-by: Paul Cercueil <paul@crapouillou.net>