lightrec: Read/write GTE registers the same way in C and in the dynarec
The GTE register cache is an array of 32 32-bit data registers, and 32
32-bit control registers. Some of these are actually accessed as signed
16-bit, some are accessed as unsigned 16-bit.

The dynarec will use 16-bit load/stores for these 16-bit registers. This
has the effect that the upper 16-bits of the 32-bit words are left
untouched.

Therefore, the interpreter should also reproduce this behaviour, instead
of sign-extending or zero-extending the 16-bit values to 32-bit,
otherwise the debugger would find differences in the GTE registers all
the time.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>