optimizer: Improve early unload registers optimization pass
Greatly improve the register liveliness analysis done in the early
unload registers optimization pass.

Instead of getting the last-read and last-write offsets of each register
at the end of the block, which will only find the last occurences where
each register is read or written, the algorithm is now analyzing across
the entire block. Additionally, instead of only adding UNLOAD flags, it
will now also add, when applicable, the new CLEAN and DISCARD flags,
which allow to clean a register without invalidating it, and unload a
register without write-back, respectively.

The benefits vs. the old implementation is that registers are unloaded
way more often, which means that host registers are way less likely to
be clobbered by emulated registers that aren't in use anymore.
Therefore, it reduces the amount of register swaps between live
registers and the register cache. Additionally, registers unloaded
before branches won't have to be invalidated in each branch-taken path.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>