disassembler: Add accessors for opcode flags
Add inline functions that can be used to retrieve the state of an opcode
flag.

The advantage of these inline functions is that they will always return
false if the corresponding optimization is not enabled, and in that
case, since they are static inline, the compiler will then be able to
detect paths that depend on the flag as dead code.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>