Jump to dispatcher directly instead of jumping to end-of-block
Until now the opcodes that caused the execution to return to the
dispatcher were branching to the end of the block, where the return
address was loaded from memory, and then jumped to.

Since the return address is no more loaded from memory, and the
end-of-block code is just a single jump opcode, we have nothing to gain
by "factorizing" the jump-to-dispatcher code, and all opcodes that cause
the execution to return to the dispatcher can now simply jump to the
dispatcher directly.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>