emitter: Avoid indirection when doing local branches
The standard generated code for a conditional local branch is to
generate an unconditional branch after a register set store-back, and
have all of it wrapped in a conditional branch that "jumps above" the
registers store-back and the unconditional branch, with the
particularity that the condition of this new branch is the opposite of
the one we are generating the code for.

For instance, let's see the generated code for a local
"BEQ r1, r2, addr":

      bne    r1, r2, end
      nop
      ...    # store-back registers
      j      addr
      nop
end:
      ...

However, under certain circumstances, there are no opcodes between the
conditional opcode and the unconditional opcode. This can be simplified
with a single conditional opcode that uses the same condition as the
opcode we're generating the code for.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>