lightrec: Fix current/target cycles not properly restored after wrapper
Both the current cycle and target cycle may be updated in a call to the
C wrapper. Therefore, we cannot cache the current cycle into a
callee-saved register.

Since state->current_cycle and state->target_cycle are both u32, the
load/store were additionally changed to work with u32. This may not have
any real effect but it's more correct.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>