regcache: Add workaround for t9 being overwritten on MIPS
On MIPS, the t9 register points (on function entry) to the address of
the function; which means that it will either be used as the target for
the JALR opcode, or overwritten in the JALR's delay slot.

Therefore, unless the temporary register used to compute the jump
address is the t9 register, we want to unload anything mapped to t9.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>