regcache: Add support for cached temporary values
Add a new priority level, REG_IS_TEMP_VALUE, which corresponds to a
temporary register that holds a constant value, for instance a bitmask
or an absolute address.

This will be useful for instance to cache the address of the C wrappers,
so that the address won't be constructed into a temporary register for
each one of several consecutive load/stores.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>