regcache: Use JIT_R0 as the cycles register
Use JIT_R0 instead of the last callee-saved register as the cycles
register.

The advantage of using a caller-saved register is that Lightning won't
generate code to save/restore its value; which means that the cycles
register can now be updated in the wrapper without Lightning getting
in the way.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>