lightrec: Remove trampoline in C wrapper
The C wrapper was previously cheesing Lightning to not generate
save/restore code for the callee-saved registers, so that the cycles
register could be updated.

Instead of incrementing the STATE register then computing the diff to
know the wrapper to use, the entry points now load the C function
pointer directly into JIT_R1 and branch to the wrapper's real prolog.

Since the STATE register is not modified anymore in the wrapper, and the
CYCLE register is caller-save (and therefore Lightning won't generate
spill/reload code for it), the trampoline trick can now be removed.

This added branch is not much of a big deal since we also removed two
branches by removing the trampoline.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>