emitter: Always use unsigned load opcodes on big-endian
On big-endian systems, the value loaded from memory will have to be
byte-swapped. Therefore, a unsigned load opcode can be used.

The benefit of using an unsigned load opcode is that on PowerPC
Lightning will be able to merge the load opcode and the endian-swap
opcode to emit a single instruction.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>