Wrap opcode list in new struct opcode_list
When allocating a 'struct opcode' list, Lightrec will now allocate a
'struct opcode_list' with the requested number of entries. This new
structure has an extra 'nb_ops' field with contains the number of
entries in the list.

Note that the lightrec_disassemble() function will still return a
'struct opcode *' pointer, and the 'opcode_list' field of the struct
block will still be a 'struct opcode *'. These actually point to the
first opcode of the struct opcode_list, so the parent struct opcode_list
is sort-of hidden. It is done this way to reduce the size of this
commit, which would be huge otherwise.

The lightrec_free_opcode_list() function will then simply retrieve the
pointer to the 'struct opcode_list' from the 'struct opcode' pointer
using the container_of() macro.

The rationale behind this change, is that this decouples the opcode list
from the block, which will later permit to free an opcode list and the
corresponding block independently. The dependence has always been a
problem with the threaded compiler, because destroying a block before
destroying an opcode list related to this block would cause
use-after-free errors.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>