lightrec: Reap blocks if block at next PC is dead
If the block at the next PC is dead but has never been compiled, then
its function pointer is NULL and we cannot execute the block.

In that case, we can just request the reaper to reap the dead blocks,
and in the next iteration of the loop, we will create a new block.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>