lightrec: Check BLOCK_IS_DEAD flag before removing outdated block
When a block is detected as outdated by the main thread, we need to
verify that it hasn't been marked as dead already before removing it. If
it has, that means the block is in the reaper's queue, and removing it
now would cause the block to be removed twice.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>