lightrec: Fix calling state->get_next_block from dispatcher
The (state->get_next_block) function pointer should be a valid jump
address that can be written into the code LUT. It allows to force the
dispatcher to run the get_next_block_func() C function, for instance
when a block needs to be recompiled, without having to write NULL to
the LUT entry, which would specify that the block is maybe outdated.

Unfortunately, the pointer was invalid, as the label referenced was
located after the code that saved the current and target cycle counters.

Fix this issue by moving the jump label before the code that saves the
current and target cycle counters.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>