lightrec: Only write LUT once in lightrec_rw_generic_cb
When an I/O opcode is tagged, the code LUT is reset at the block's
PC so that the next time the block is executed, the dispatcher will jump
to the C routine and recompile the block.

However - the code LUT only has to be written the first time; if the
"block must recompile" flag is already present, there is no need to
write the code LUT once again.

Note that writing NULL to the code LUT here may race with the threaded
compiler, which might have set just before the LUT entry for a block
that encompasses ours. This is not a problem, because:
- if the new block's entry point overwrites our NULL pointer, we won't
  be able to recompile our block, which is fine, since it will be
  replaced;
- if our NULL pointer overwrites the new block's entry point, then the
  current block will recompile, and it will simply coexist with the new
  block. This is fine, and it should be a really rare race anyway.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>